requirements.txt
Django==5.2.4
psycopg2-binary
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
PyJWT==2.9.0
sqlparse
asgiref
gunicorn
dj-database-url
whitenoise
django-import-export
pytz==2023.3
